refactor: changes files dir
