+++ src/components/partials/circle-progress/interfaces.ts
