+++ src/components/partials/circle-progress/styles.ts
