+++ src/components/partials/index.ts
@@ -1,2 +1,2 @@
 export * from "./materialListItem/";
-export * from "./chart/circle-progress/";
+export * from "./circle-progress";
